security Filter

+++ fast-security/src/main/java/com/fast/security/JdbcTokenRepository.java
@@ -2,13 +2,12 @@ package com.fast.security;
 
 import org.springframework.security.web.authentication.rememberme.JdbcTokenRepositoryImpl;
 
-
 public class JdbcTokenRepository extends JdbcTokenRepositoryImpl {
 
-    public Long queryTokenCount(String username) {
+    public int queryTokenCount(String username) {
         String sql = "select count(*) from persistent_logins where username=?";
         Object[] object = {username};
-        return getJdbcTemplate().queryForObject(sql, object, Long.class);
+        return getJdbcTemplate().queryForObject(sql, object, int.class);
     }
 
 }
